requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.9
PyMySQL
Flask-Heroku
gunicorn
psycopg2
python-decouple==3.1
Flask-Mail
SQLAlchemy-Utils
cryptography
flask_cors